pyproject.toml
[tool.poetry]
name = "apachekafka-demo"
version = "0.1.0"
description = ""
authors = ["Arshid Aslam <arshidbaba93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
